docs(recipe): remove TableCard recipe (#1538)
- this recipe, as-is, is tightly coupled to one particular project and can't be repurposed
- move it to the one page in storybook where it is used for now
