+++ .storybook/pages/CoursePlannerEdit/CoursePlannerEdit.module.css
@@ -1,5 +1,6 @@
+@import '../../../src/design-tokens/mixins.css';
 /*-----------------------------------------*\
-    # SPACING COURSE PLANNER EDIT CLASSES
+    # COURSE PLANNER EDIT CLASSES
 \*-----------------------------------------*/
 
 /**
@@ -7,10 +8,42 @@
  * 
  * Forces margin top to 2rem (on smaller viewports only).
  */
- .button--mobile {
-    margin-top: var(--eds-size-4) !important;
-  
+.button--mobile {
+  margin-top: var(--eds-size-4) !important;
+
+  @media all and (min-width: $eds-bp-md) {
+    margin-top: unset !important;
+  }
+}
+
+/*------------------------------------*\
+    # TABLE CARD
+\*------------------------------------*/
+
+/**
+ * Card with a title, table, and button within
+ */
+
+/**
+* Table card table
+*
+* Actual table within the card.
+*/
+.table-card__table {
+  color: var(--eds-theme-color-text-neutral-subtle);
+  margin-bottom: var(--eds-size-2);
+}
+
+/**
+* Table card table header cell
+*/
+.table-card__table-header-cell {
+  /**
+  * First child within table card table header cell
+  */
+  &:first-child {
     @media all and (min-width: $eds-bp-md) {
-      margin-top: unset !important;
+      width: 11.875rem;
     }
   }
+}
