+++ .storybook/pages/CoursePlannerEdit/CoursePlannerEdit.tsx
@@ -1,4 +1,6 @@
+import clsx from 'clsx';
 import React, { useState } from 'react';
+import type { ReactNode } from 'react';
 
 import ProjectCard from './ProjectCard';
 
@@ -22,16 +24,27 @@ import {
   GridItem,
   Text,
   DropdownMenuItem,
+  Card,
+  CardBody,
+  NumberIcon,
+  TableHeader,
+  TableRow,
+  Table,
+  TableBody,
+  TableCell,
+  TableHeaderCell,
+  CardFooter,
 } from '../../../src';
 
 import type { NewState } from '../../../src/components/DragDrop/DragDrop';
 
 import CardWithNotification from '../../recipes/CardWithNotification';
-import TableCard from '../../recipes/TableCard';
+import NumberIconList from '../../recipes/NumberIconList';
 
 import EmptyImage from '../../static/hand-pencil.svg';
 import styles from './CoursePlannerEdit.module.css';
 
+NumberIconList;
 const CognitiveSkillColumns = [
   {
     title: 'Least covered Cognitive Skills',
@@ -263,6 +276,106 @@ const StandardsRows = [
     ],
   },
 ];
+
+interface Item {
+  'aria-label': string;
+  complete: boolean;
+}
+
+interface Row {
+  value1: string;
+  projects: Item[];
+}
+
+interface Column {
+  title: string;
+}
+
+interface CardProps {
+  /**
+   * CSS class names that can be appended to the component.
+   */
+  className?: string;
+  title?: string;
+  buttonContent?: ReactNode;
+  tableColumns: Column[];
+  tableRows: Row[];
+}
+
+/**
+ * A Card containing a Table.
+ */
+export const TableCard = ({
+  className,
+  title,
+  buttonContent,
+  tableRows,
+  tableColumns,
+  ...other
+}: CardProps) => {
+  const componentClassName = clsx(styles['table-card'], className);
+
+  return (
+    <Card className={componentClassName} {...other}>
+      <CardBody>
+        <Heading
+          as="h2"
+          className="mb-4"
+          size="title-sm"
+          variant="neutral-strong"
+        >
+          {title}
+        </Heading>
+        <Table className={styles['table-card__table']} title={title}>
+          <TableHeader>
+            <TableRow variant="header">
+              {tableColumns.map((item, index) => {
+                return (
+                  <TableHeaderCell
+                    className={styles['table-card__table-header-cell']}
+                    key={'table-header-cell-' + item.title}
+                  >
+                    {item.title}
+                  </TableHeaderCell>
+                );
+              })}
+            </TableRow>
+          </TableHeader>
+          <TableBody>
+            {tableRows.map((item) => {
+              return (
+                <TableRow key={`table-row-${item.value1}`}>
+                  <TableHeaderCell variant="body">
+                    {item.value1}
+                  </TableHeaderCell>
+                  <TableCell>
+                    <NumberIconList>
+                      {item.projects.map((item, index) => {
+                        return (
+                          <NumberIcon
+                            aria-label={item['aria-label']}
+                            incomplete={!item.complete}
+                            key={`number-icon-${item['aria-label']}`}
+                            number={index + 1}
+                            numberIconTitle={`incomplete step ${index + 1}`}
+                            size="sm"
+                          />
+                        );
+                      })}
+                    </NumberIconList>
+                  </TableCell>
+                </TableRow>
+              );
+            })}
+          </TableBody>
+        </Table>
+      </CardBody>
+      <CardFooter>
+        <Button status="neutral">{buttonContent}</Button>
+      </CardFooter>
+    </Card>
+  );
+};
 
 export const CoursePlannerEdit = () => {
   const container1EmptyContent = () => {
