--- .storybook/recipes/TableCard/TableCard.module.css
@@ -1,33 +1,0 @@
-@import '../../../src/design-tokens/mixins.css';
-
-/*------------------------------------*\
-    # TABLE CARD
-\*------------------------------------*/
-
-/**
- * Card with a title, table, and button within
- */
-
-/**
-* Table card table
-*
-* Actual table within the card.
-*/
-.table-card__table {
-  color: var(--eds-theme-color-text-neutral-subtle);
-  margin-bottom: var(--eds-size-2);
-}
-
-/**
-* Table card table header cell
-*/
-.table-card__table-header-cell {
-  /**
-  * First child within table card table header cell
-  */
-  &:first-child {
-    @media all and (min-width: $eds-bp-md) {
-      width: 11.875rem;
-    }
-  }
-}
