--- .storybook/recipes/TableCard/index.ts
@@ -1,1 +1,0 @@
-export { TableCard as default } from './TableCard';
